Add files via upload
First, I formatted the table to improve readability and froze the headers of the table so that they are always on display, no matter where you are looking in the table. Then I loaded the table into Power Query.
In Power Query I ensure the data was the appropriate data type and looked for any errors or nulls in the data. There were none in this data. I loaded this back into Excel and then checked for duplicate rows of which there were none. 
On “Processed Data (2)” I calculated the descriptive statistics at the bottom of the table including the Total, Mean, Minimum, Maximum, Standard Deviation and the Variance of the sample. 
Once I was satisfied the data was clean I the uploaded it onto the Data Model.
From here I created some charts in Power Pivot, added graphics, slicers and timeline and moved these to create an interactive dashboard.